changed the file structure for seller part
