+++ backend/controllers/authControllerSeller.js
