+++ backend/models/seller.js
@@ -83,4 +83,4 @@ userSchema.methods.getResetPasswordToken = function () {
 
 }
 
-module.exports = mongoose.model('User', userSchema);
+module.exports = mongoose.model('Seller', userSchema);
