+++ backend/routes/orderSeller.js
@@ -9,17 +9,17 @@ const {
     updateOrder,
     deleteOrder
 
-} = require('../controllers/orderController')
+} = require('../controllers/orderControllerSeller')
 
-const { isAuthenticatedUser, authorizeRoles } = require('../middlewares/auth')
+const { isAuthenticatedUser, authorizeRoles } = require('../middlewares/authSeller')
 
 router.route('/order/new').post(isAuthenticatedUser, newOrder);
 
 router.route('/order/:id').get(isAuthenticatedUser, getSingleOrder);
 router.route('/orders/me').get(isAuthenticatedUser, myOrders);
 
-router.route('/admin/orders/').get(isAuthenticatedUser, authorizeRoles('seller'), allOrders);
-router.route('/admin/order/:id')
+router.route('/seller/orders/').get(isAuthenticatedUser, authorizeRoles('seller'), allOrders);
+router.route('/seller/order/:id')
     .put(isAuthenticatedUser, authorizeRoles('seller'), updateOrder)
     .delete(isAuthenticatedUser, authorizeRoles('seller'), deleteOrder);
