+++ frontend/src/actionsSeller/orderActionsSeller.js
@@ -90,13 +90,13 @@ export const getOrderDetails = (id) => async (dispatch) => {
     }
 }
 
-// Get all orders - ADMIN
+// Get all orders - seller
 export const allOrders = () => async (dispatch) => {
     try {
 
         dispatch({ type: ALL_ORDERS_REQUEST });
 
-        const { data } = await axios.get(`/api/v1/admin/orders`)
+        const { data } = await axios.get(`/api/v1/seller/orders`)
 
         dispatch({
             type: ALL_ORDERS_SUCCESS,
@@ -123,7 +123,7 @@ export const updateOrder = (id, orderData) => async (dispatch) => {
             }
         }
 
-        const { data } = await axios.put(`/api/v1/admin/order/${id}`, orderData, config)
+        const { data } = await axios.put(`/api/v1/seller/order/${id}`, orderData, config)
 
         dispatch({
             type: UPDATE_ORDER_SUCCESS,
@@ -144,7 +144,7 @@ export const deleteOrder = (id) => async (dispatch) => {
 
         dispatch({ type: DELETE_ORDER_REQUEST })
 
-        const { data } = await axios.delete(`/api/v1/admin/order/${id}`)
+        const { data } = await axios.delete(`/api/v1/seller/order/${id}`)
 
         dispatch({
             type: DELETE_ORDER_SUCCESS,
