+++ frontend/src/components/layout/Header.js
@@ -75,7 +75,7 @@ const Header = () => {
                 {user && user.role === 'seller' && (
                   <Link className="dropdown-item" to="/dashboard">Dashboard</Link>
                 )}
-                <Link className="dropdown-item" to="/orders/me">Orders</Link>
+                <Link className="dropdown-item" to="/seller/orders">Orders</Link>
                 <Link className="dropdown-item" to="/me">Profile</Link>
                 <Link className="dropdown-item text-danger" to="/" onClick={logoutHandler}>
                   Logout
