+++ frontend/src/components/route/ProtectedRouteSeller.js
