+++ frontend/src/components/seller/Dashboard.js
@@ -8,7 +8,7 @@ import Sidebar from './Sidebar'
 import { useDispatch, useSelector } from 'react-redux'
 
 import { getSellerProducts } from '../../actions/productActions'
-import { allOrders } from '../../actions/orderActions'
+import { allOrders } from '../../actions/orderActionsSeller'
 import { allUsers } from '../../actions/userActions'
 
 const Dashboard = () => {
@@ -58,7 +58,7 @@ const Dashboard = () => {
                             </div>
 
                             <div className="row pr-4">
-                                <div className="col-xl-3 col-sm-6 mb-3">
+                                <div className="col-xl-4 col-sm-6 mb-3">
                                     <div className="card text-white bg-success o-hidden h-100">
                                         <div className="card-body">
                                             <div className="text-center card-font-size">Products<br /> <b>{products && products.length}</b></div>
@@ -73,12 +73,12 @@ const Dashboard = () => {
                                 </div>
 
 
-                                <div className="col-xl-3 col-sm-6 mb-3">
+                                <div className="col-xl-4 col-sm-6 mb-3">
                                     <div className="card text-white bg-danger o-hidden h-100">
                                         <div className="card-body">
                                             <div className="text-center card-font-size">Orders<br /> <b>{orders && orders.length}</b></div>
                                         </div>
-                                        <Link className="card-footer text-white clearfix small z-1" to="/admin/orders">
+                                        <Link className="card-footer text-white clearfix small z-1" to="/seller/orders">
                                             <span className="float-left">View Details</span>
                                             <span className="float-right">
                                                 <i className="fa fa-angle-right"></i>
@@ -88,7 +88,7 @@ const Dashboard = () => {
                                 </div>
 
 
-                                <div className="col-xl-3 col-sm-6 mb-3">
+                                {/* <div className="col-xl-3 col-sm-6 mb-3">
                                     <div className="card text-white bg-info o-hidden h-100">
                                         <div className="card-body">
                                             <div className="text-center card-font-size">Users<br /> <b>{users && users.length}</b></div>
@@ -100,10 +100,10 @@ const Dashboard = () => {
                                             </span>
                                         </Link>
                                     </div>
-                                </div>
+                                </div> */}
 
 
-                                <div className="col-xl-3 col-sm-6 mb-3">
+                                <div className="col-xl-4 col-sm-6 mb-3">
                                     <div className="card text-white bg-warning o-hidden h-100">
                                         <div className="card-body">
                                             <div className="text-center card-font-size">Out of Stock<br /> <b>{outOfStock}</b></div>
