+++ frontend/src/components/seller/Sidebar.js
@@ -25,15 +25,13 @@ const Sidebar = () => {
                     </li>
 
                     <li>
-                        <Link to="/admin/orders"><i className="fa fa-shopping-basket"></i> Orders</Link>
+                        <Link to="/seller/orders"><i className="fa fa-shopping-basket"></i> Orders</Link>
                     </li>
 
-                    <li>
-                        <Link to="/admin/users"><i className="fa fa-users"></i> Users</Link>
-                    </li>
+                
 
                     <li>
-                        <Link to="/admin/reviews"><i className="fa fa-star"></i> Reviews</Link>
+                        <Link to="/seller/reviews"><i className="fa fa-star"></i> Reviews</Link>
                     </li>
 
                 </ul>
