+++ frontend/src/components/sellerProfile/NewPassword.js
