+++ frontend/src/components/sellerProfile/Profile.js
@@ -37,7 +37,7 @@ const Profile = () => {
                             <p>{String(user.createdAt).substring(0, 10)}</p>
 
                             {user.role !== 'admin' && (
-                                <Link to="/orders/me" className="btn btn-danger btn-block mt-5">
+                                <Link to="/seller/orders" className="btn btn-danger btn-block mt-5">
                                     My Orders
                                 </Link>
                             )}
