+++ frontend/src/reducersSeller/orderReducers.js
@@ -20,7 +20,7 @@ import {
     ORDER_DETAILS_SUCCESS,
     ORDER_DETAILS_FAIL,
     CLEAR_ERRORS
-} from '../constants/orderConstants'
+} from '../constantsSeller/orderConstants'
 
 export const newOrderReducer = (state = {}, action) => {
     switch (action.type) {
