+++ frontend/src/reducersSeller/userReducers.js
@@ -39,7 +39,7 @@ import {
     LOGOUT_SUCCESS,
     LOGOUT_FAIL,
     CLEAR_ERRORS
-} from '../constants/userConstants'
+} from '../constantsSeller/userConstants'
 
 export const authReducer = (state = { user: {} }, action) => {
     switch (action.type) {
